break: rename CancelActionMarker to cancel
