+++ packages/dataparcels-docs/src/pages/api/cancel.jsx
@@ -2,13 +2,13 @@
 import React from 'react';
 import Layout from 'layout/Layout';
 import ContentNav from 'shape/ContentNav';
-import CancelActionMarkerMarkdown from 'pages/api/CancelActionMarker.mdx';
+import cancelMarkdown from 'pages/api/cancel.mdx';
 
 export default () => <Layout>
     <ContentNav
-        content={() => <CancelActionMarkerMarkdown />}
+        content={() => <cancelMarkdown />}
         pageNav={[
-            '# CancelActionMarker'
+            '# cancel'
         ]}
     />
 </Layout>;
