--- packages/dataparcels/CancelActionMarker.js
@@ -1,2 +1,0 @@
-/* eslint-disable */
-module.exports = require('./lib/change/CancelActionMarker.js').default;
