+++ packages/dataparcels/src/change/ChangeRequestReducer.js
@@ -13,7 +13,7 @@ import pipeWith from 'unmutable/lib/util/pipeWith';
 import composeWith from 'unmutable/lib/util/composeWith';
 import {ReducerInvalidActionError} from '../errors/Errors';
 import {ReducerInvalidStepError} from '../errors/Errors';
-import {isCancelledError} from './CancelActionMarker';
+import {isCancelledError} from './cancel';
 
 import del from '../parcelData/delete';
 import deleteSelfWithMarker from '../parcelData/deleteSelfWithMarker';
