--- packages/react-dataparcels/CancelActionMarker.js
@@ -1,2 +1,0 @@
-/* eslint-disable */
-module.exports = require('dataparcels/CancelActionMarker.js');
