+++ packages/react-dataparcels/cancel.js
@@ -1,0 +1,2 @@
+/* eslint-disable */
+module.exports = require('dataparcels/cancel.js');
